+++ app/src/main/kotlin/com/wisp/app/ui/component/ProfileQrSheet.kt
@@ -19,6 +19,7 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ContentCopy
 import androidx.compose.material.icons.outlined.CurrencyBitcoin
 import androidx.compose.material.icons.outlined.Person
+import androidx.compose.material.icons.outlined.QrCodeScanner
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -29,8 +30,11 @@ import androidx.compose.material3.TabRow
 import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -47,12 +51,13 @@ import coil3.compose.AsyncImage
 import com.wisp.app.R
 import com.wisp.app.nostr.Nip19
 import com.wisp.app.nostr.hexToByteArray
+import com.wisp.app.toRoute
 import com.wisp.app.ui.theme.WispThemeColors
 import kotlinx.coroutines.launch
 
 /**
- * Consolidated QR sheet showing Nostr identity and Lightning address
- * in a tabbed bottom sheet with swipeable pages.
+ * Consolidated QR sheet showing Nostr identity, Lightning address, and a scan
+ * tab in a tabbed bottom sheet with swipeable pages.
  */
 @OptIn(ExperimentalMaterial3Api::class, androidx.compose.foundation.ExperimentalFoundationApi::class)
 @Composable
@@ -60,6 +65,7 @@ fun ProfileQrSheet(
     pubkeyHex: String,
     avatarUrl: String? = null,
     lud16: String? = null,
+    onNavigate: (String) -> Unit = {},
     onDismiss: () -> Unit
 ) {
     val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
@@ -73,8 +79,12 @@ fun ProfileQrSheet(
     val lightningQr = remember(lud16) { lud16?.let { generateQrBitmap(it) } }
 
     val hasLightning = lud16 != null
-    val pageCount = if (hasLightning) 2 else 1
+    // Page indices: 0 = Nostr, 1 = Lightning (if present), last = Scan.
+    val lightningPage = if (hasLightning) 1 else -1
+    val scanPage = if (hasLightning) 2 else 1
+    val pageCount = if (hasLightning) 3 else 2
     val pagerState = rememberPagerState(pageCount = { pageCount })
+    var scanError by remember { mutableStateOf<String?>(null) }
 
     ModalBottomSheet(
         onDismissRequest = onDismiss,
@@ -86,23 +96,22 @@ fun ProfileQrSheet(
                 .fillMaxWidth()
                 .padding(bottom = 32.dp)
         ) {
-            // Tabs
-            if (hasLightning) {
-                TabRow(
-                    selectedTabIndex = pagerState.currentPage,
-                    containerColor = MaterialTheme.colorScheme.surface,
-                    contentColor = MaterialTheme.colorScheme.primary,
-                    divider = {}
-                ) {
+            TabRow(
+                selectedTabIndex = pagerState.currentPage,
+                containerColor = MaterialTheme.colorScheme.surface,
+                contentColor = MaterialTheme.colorScheme.primary,
+                divider = {}
+            ) {
+                Tab(
+                    selected = pagerState.currentPage == 0,
+                    onClick = { scope.launch { pagerState.animateScrollToPage(0) } },
+                    text = { Text("Nostr") },
+                    icon = { Icon(Icons.Outlined.Person, null, modifier = Modifier.size(18.dp)) }
+                )
+                if (hasLightning) {
                     Tab(
-                        selected = pagerState.currentPage == 0,
-                        onClick = { scope.launch { pagerState.animateScrollToPage(0) } },
-                        text = { Text("Nostr") },
-                        icon = { Icon(Icons.Outlined.Person, null, modifier = Modifier.size(18.dp)) }
-                    )
-                    Tab(
-                        selected = pagerState.currentPage == 1,
-                        onClick = { scope.launch { pagerState.animateScrollToPage(1) } },
+                        selected = pagerState.currentPage == lightningPage,
+                        onClick = { scope.launch { pagerState.animateScrollToPage(lightningPage) } },
                         text = { Text("Lightning") },
                         icon = {
                             if (useZapBolt) {
@@ -113,11 +122,16 @@ fun ProfileQrSheet(
                         }
                     )
                 }
+                Tab(
+                    selected = pagerState.currentPage == scanPage,
+                    onClick = { scope.launch { pagerState.animateScrollToPage(scanPage) } },
+                    text = { Text("Scan") },
+                    icon = { Icon(Icons.Outlined.QrCodeScanner, null, modifier = Modifier.size(18.dp)) }
+                )
             }
 
             Spacer(Modifier.height(16.dp))
 
-            // Pages
             HorizontalPager(
                 state = pagerState,
                 modifier = Modifier.fillMaxWidth()
@@ -181,8 +195,7 @@ fun ProfileQrSheet(
                                 clipboardManager.setText(AnnotatedString(npub))
                             })
                         }
-                        1 -> {
-                            // Lightning QR
+                        lightningPage -> {
                             if (lightningQr != null && lud16 != null) {
                                 Box(
                                     contentAlignment = Alignment.Center,
@@ -234,6 +247,29 @@ fun ProfileQrSheet(
                                 CopyableRow(text = lud16, onCopy = {
                                     clipboardManager.setText(AnnotatedString(lud16))
                                 })
+                            }
+                        }
+                        scanPage -> {
+                            Box(
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                                    .height(360.dp)
+                            ) {
+                                QrScanner(
+                                    onResult = { raw ->
+                                        val route = Nip19.decodeNostrQr(raw)?.toRoute()
+                                        if (route != null) {
+                                            scanError = null
+                                            onNavigate(route)
+                                            onDismiss()
+                                        } else {
+                                            scanError = "Not a Nostr QR code"
+                                        }
+                                    },
+                                    modifier = Modifier.fillMaxWidth(),
+                                    promptText = scanError
+                                        ?: "Point at an npub, note, profile, event, or address QR"
+                                )
                             }
                         }
                     }
